Added Timestep control
- Added physics time step control through lua
- Began tweaking precision
- Began prototyping collision triggers

+++ Game/EventReceiver.cpp
@@ -5,7 +5,7 @@ using namespace FCE;
 
 FEventReceiver::FEventReceiver()
 {
-	for(int i = 0; i < KEY_KEY_CODES_COUNT; i++)
+	for(u16 i = 0; i < KEY_KEY_CODES_COUNT; i++)
 	{
 		keys[i] = false;
 	}
